+++ axum-on-rails/src/lib.rs
@@ -3,6 +3,6 @@ use bb8_postgres::PostgresConnectionManager;
 use tokio_postgres::NoTls;
 
 pub mod cli;
-pub mod test_helpers;
+pub mod test;
 
 pub type ConnectionPool = Pool<PostgresConnectionManager<NoTls>>;
